pyproject.toml
[tool.poetry]
name = "schema-viewer"
version = "0.1.0"
description = ""
authors = ["statifc0x <44530786+staticf0x@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Jinja2 = "^3.1.1"
jsonref = "^0.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
